Add files via upload
Shredder blades are used in machines that cut various materials into smaller pieces, including paper, plastic, wood, and metal. They are a crucial component in shredders, which are used for recycling, waste disposal, document destruction, and other applications. As per the reengineering design model has been developed for a client through Solidworks to develop a CNC programming for Mastercam.